Add support for SimpleAlarmClock/com.better.alarm.
This adds support for alarms from SimpleAlarmClock, AKA
com.better.alarm.

This is available from the Play store at
https://play.google.com/store/apps/details?id=com.better.alarm and from GitHub at
https://github.com/yuriykulikov/AlarmClock

SimpleAlarmClock doesn't do notifications of upcoming alarms, and
doesn't do timers.  But it does do alarms, and unlike the Google Clock
app on Oreo, it does alarms with notifications with usable actions.

It is also FOSS (Apache 2.0 license), and was recommended by
u/ozdreaming on Reddit.  Good enough to make me pick it.