Create a foreground notification.
This creates a foreground notification for the service.

We do our best not to be noisy about it, but at the end of the day we
need one of these if we want to work on Oreo at all.

Right now, clicking on it does nothing.  At some point we should
probably make it bring up the Alarmist settings.